+++ crates/veryl/src/cmd_new.rs
@@ -26,7 +26,7 @@ impl CmdNew {
             write!(file, "{}", toml).into_diagnostic()?;
             file.flush().into_diagnostic()?;
 
-            println!("Created \"{}\" package", name.to_string_lossy());
+            println!("Created \"{}\" project", name.to_string_lossy());
         } else {
             bail!("path \"{}\" is not valid", self.opt.path.to_string_lossy());
         }